--- Assets/EkaterinaFolder/Enemy.cs
@@ -1,18 +1,0 @@
-using UnityEngine;
-
-    public class Enemy : MonoBehaviour
-    {
-        // Start is called before the first frame update
-        void Start()
-        {
-
-        }
-
-        // Update is called once per frame
-        void Update()
-        {
-            
-        }
-    }
-
-
