--- Assets/EkaterinaFolder/Player.cs
@@ -1,19 +1,0 @@
-using UnityEngine;
-
-    public class Player : MonoBehaviour
-    {
-        
-        // Start is called before the first frame update
-        void Start()
-        {
-
-        }
-
-        // Update is called once per frame
-        void Update()
-        {
-            
-        }
-    }
-
-
