--- Assets/EkaterinaFolder/SceneLoader.cs
@@ -1,13 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class SceneLoader : MonoBehaviour
-{
-    public void LoadScene(string scene)
-    {
-        SceneManager.LoadScene(scene);
-    }
-
-}
