+++ Assets/EkaterinaFolder/Scenes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c463b9a226e4e4a369b47d6e85b8c6b7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
