+++ Assets/EkaterinaFolder/Scenes/WerewolfReviversGame.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 15bc0d53c5d98420d9c8b4c5328b987f
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
