+++ Assets/EkaterinaFolder/Scenes/WerewolfReviversLose.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 06f12f333132b4b06996776afd59decb
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
