+++ Assets/EkaterinaFolder/Scenes/WerewolfReviversTitle.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b88cbd48a781042628b22a033bf92902
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
