+++ Assets/EkaterinaFolder/Scenes/WerewolfReviversWin.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4f996bf6cb89a4cc7aa0605c08e9696c
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
