+++ Assets/EkaterinaFolder/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fa3c83d937e6b417d9b77b85dc028b21
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
