+++ Assets/EkaterinaFolder/Scripts/Enemy.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: cf52f5abd6712d8438d55d6f784b42db
+guid: e5fe558db619343928759c27b75e013d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
