+++ Assets/EkaterinaFolder/Scripts/GameController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3dc5bbbd5c0b0394688255578da64290
+guid: 934d84dd063764dc8b9834e0e5d8e93b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
