+++ Assets/EkaterinaFolder/Scripts/Player.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ad72090fe474af542a54d6305312888e
+guid: 0e81f62d124a84a4fa18ccf210acc68f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
