+++ Assets/EkaterinaFolder/Scripts/SceneLoader.cs
@@ -1,0 +1,20 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+namespace EkaterinaVampireSurvivors
+{
+
+    public class SceneLoader : MonoBehaviour
+    {
+        public void LoadScene(string scene)
+        {
+            SceneManager.LoadScene(scene);
+        }
+
+    }
+
+}
+
+
