+++ Assets/EkaterinaFolder/Scripts/SceneLoader.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0e8e6271bc392de499d82587276195f4
+guid: e2bf90a75f3f0489dbd83cbdbda74407
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
