+++ Assets/EkaterinaFolder/Scripts/Spawner.cs
@@ -1,0 +1,11 @@
+using UnityEngine;
+
+namespace EkaterinaVampireSurvivors
+{
+    public class Spawner : MonoBehaviour
+    {
+
+    }
+}
+
+
