+++ Assets/EkaterinaFolder/Scripts/Spawner.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a44475117f2825e448da7ccbbd635955
+guid: 518148e1cf562465bb9ed9044e088612
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
