--- Assets/EkaterinaFolder/Spawner.cs
@@ -1,6 +1,0 @@
-using UnityEngine;
-
-public class Spawner : MonoBehaviour
-{
-
-}
