+++ UserSettings/EditorUserSettings.asset
@@ -11,6 +11,12 @@ EditorUserSettings:
     RecentlyUsedSceneGuid-1:
       value: 0104575003560b095c5a597543720f44464e197d7d2a7561797e4b35e3b46069
       flags: 0
+    RecentlyUsedSceneGuid-2:
+      value: 5307525757535a085e5f5f7012775d40404e417c7b7f70312b2d486ae6e5373a
+      flags: 0
+    RecentlyUsedSceneGuid-3:
+      value: 5204560555515c080e5b087b48770d40124e1b722e7c25657e7b4531bbb8633e
+      flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
       flags: 0
